redesign: update wallet connection dialogs
- Put EVM wallet connection through wagmi under a feature flag
- Remove redundant ModalSelectWallet dialog
