+++ .env.development
@@ -18,6 +18,7 @@ NEXT_PUBLIC_LINK_X="https://x.com/DefuseProtocol"
 NEXT_PUBLIC_LINK_DISCORD="https://discord.gg/rdGAgDRs"
 NEXT_PUBLIC_LINK_DOCS="https://docs.google.com/document/d/1yK5zjZpdA8IjiGc3JSzb7dfyt-5qsBjMMMi4KbYNHkM/edit?usp=sharing"
 NEXT_PUBLIC_DEV_MODE=false
+NEXT_PUBLIC_TURN_OFF_EVM_WALLETS=true
 
 ######################################################
 # Networks ID                                        #
