+++ .env.production
@@ -13,7 +13,7 @@ NEXT_PUBLIC_PUBLIC_MAIL="defuse@defuse.org"
 NEXT_PUBLIC_LINK_X="https://x.com/DefuseProtocol"
 NEXT_PUBLIC_LINK_DISCORD="https://discord.gg/rdGAgDRs"
 NEXT_PUBLIC_LINK_DOCS="https://docs.google.com/document/d/1yK5zjZpdA8IjiGc3JSzb7dfyt-5qsBjMMMi4KbYNHkM/edit?usp=sharing"
-
+NEXT_PUBLIC_TURN_OFF_EVM_WALLETS=true
 
 ######################################################
 # Networks ID                                        #
